chore: Service priority should be an interface
Service.Priority should be an interface so it can be omitted if not set.

Closes #3426

+++ pkg/didcomm/protocol/outofband/service_test.go
@@ -1027,7 +1027,7 @@ func TestChooseTarget(t *testing.T) {
 		expected := map[string]interface{}{
 			"id":              uuid.New().String(),
 			"type":            "did-communication",
-			"priority":        uint(0),
+			"priority":        0,
 			"recipientKeys":   []string{"my ver key"},
 			"serviceEndpoint": commonmodel.NewDIDCommV1Endpoint("my service endpoint"),
 			"RoutingKeys":     []string{"my routing key"},
